requirements.txt
djangorestframework==3.9.0
django-crispy-forms==1.7.2
gunicorn==19.7.0
pillow==5.3.0
whitenoise==3.3.1
dj-database-url==0.5.0
django==2.1.1
psycopg2==2.7.5
razorpay==1.1.1